requirements.txt
setuptools
wheel
fastapi==0.104.1
pydantic==2.5.0
uvicorn[standard]==0.24.0
requests==2.31.0
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<1.25.0
opencv-python-headless>=4.8.0,<5.0.0
deepface>=0.0.79
scipy>=1.11.0,<1.12.0
pandas>=2.1.0,<2.2.0
mtcnn>=0.1.1
scikit-learn>=1.3.0,<1.4.0
tensorflow>=2.13.0,<2.16.0
python-multipart>=0.0.6
apscheduler